+++ src/app/models/key-models/sortedsetkey.h
@@ -4,7 +4,8 @@
 class SortedSetKeyModel : public KeyModel<QPair<QByteArray, double>>
 {    
 public:
-    SortedSetKeyModel(QSharedPointer<RedisClient::Connection> connection, QString fullPath, int dbIndex, int ttl);
+    SortedSetKeyModel(QSharedPointer<RedisClient::Connection> connection,
+                      QString fullPath, int dbIndex, long long ttl);
 
     QString getType() override;
     QStringList getColumnNames() override;
